chore: update BankAccount to use BigDecimal for balance instead of double

+++ chapter-02/src/main/java/expert/os/books/architecture/patterns/scenario03/BankAccountBuilder.java
@@ -1,15 +1,17 @@
 package expert.os.books.architecture.patterns.scenario03;
 
+import java.math.BigDecimal;
+
 public class BankAccountBuilder {
     private String accountNumber;
-    private double balance;
+    private BigDecimal balance;
 
     public BankAccountBuilder accountNumber(String accountNumber) {
         this.accountNumber = accountNumber;
         return this;
     }
 
-    public BankAccountBuilder balance(double balance) {
+    public BankAccountBuilder balance(BigDecimal balance) {
         this.balance = balance;
         return this;
     }
